+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Bitmovin Player: New quality on quality change events was tracked one sample too early
 - IVS Player: Wrong initialization order on startup
 - IVS Player: Wrong order of releasing of resources when collector is detached
+- IVS Player: Reporting of negative droppedFrames in certain edgecases
 - All players: Reporting of `videostartup_time = 0` on certain edgecases where startup was very fast
 
 ## v2.15.0